feat: added more tests
