+++ __tests__/core/utils/format-duration.test.ts
@@ -1,7 +1,7 @@
 // eslint-disable-next-line import/no-unused-modules
 import { describe, expect, it } from "vitest";
 
-import formatDuration from "../../src/core/utils/format-duration";
+import formatDuration from "../../../src/core/utils/format-duration";
 
 describe("formatDuration", () => {
     it("should handle negative duration", () => {
